+++ equalsverifier-core/src/test/java/nl/jqno/equalsverifier/integration/operational/WithPrefabValuesForFieldTest.java
@@ -111,6 +111,13 @@ class WithPrefabValuesForFieldTest {
                 .verify();
     }
 
+    void succeed_whenClassHasStringPrecondition_givenPrefabValueForField() {
+        EqualsVerifier
+                .forClass(StringPrecondition.class)
+                .withPrefabValuesForField("s", "precondition:red", "precondition:blue")
+                .verify();
+    }
+
     @Test
     void throw_whenFieldDoesNotExistInClass() {
         ExpectedException
@@ -282,6 +289,32 @@ class WithPrefabValuesForFieldTest {
             if (y < 500 || y > 600) {
                 throw new IllegalArgumentException("y must be between 500 and 600! But was " + y);
             }
+        }
+    }
+
+    static final class StringPrecondition {
+
+        private final String s;
+
+        public StringPrecondition(String s) {
+            this.s = s;
+        }
+
+        @Override
+        public boolean equals(Object obj) {
+            if (s != null && !s.startsWith("precondition:")) {
+                throw new IllegalArgumentException("bad precondition: " + s);
+            }
+            if (!(obj instanceof StringPrecondition)) {
+                return false;
+            }
+            StringPrecondition other = (StringPrecondition) obj;
+            return Objects.equals(s, other.s);
+        }
+
+        @Override
+        public int hashCode() {
+            return Objects.hash(s);
         }
     }