+++ CHANGELOG.md
@@ -32,6 +32,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `forPackage(...).except(Class<?>... types)`, which was deprecated since version 3.19.
 - `forPackage(...).except(Predicate<Class<?>>... exclusionPredicate)`, which was deprecated since version 3.19.
 
+## [3.19.1] - 2025-02-17
+
+### Added
+
+- Prefab value for `java.util.concurrent.locks.ReentrantLock`. ([Issue 981](https://github.com/jqno/equalsverifier/issues/981))
+
+### Fixed
+
+- Bug in `withPrefabValueForField` where in some cases the prefab value is not used.
+
 ## [3.19] - 2025-02-06
 
 ### Added
@@ -1343,7 +1353,8 @@ Please don't use version 1.3; [it's a broken release](https://jqno.nl/post/2013/
 
 You can now use EqualsVerifier!
 
-[unreleased]: https://github.com/jqno/equalsverifier/compare/equalsverifier-3.19...HEAD
+[unreleased]: https://github.com/jqno/equalsverifier/compare/equalsverifier-3.19.1...HEAD
+[3.19.1]: https://github.com/jqno/equalsverifier/compare/equalsverifier-3.19...equalsverifier-3.19.1
 [3.19]: https://github.com/jqno/equalsverifier/compare/equalsverifier-3.18.2...equalsverifier-3.19
 [3.18.2]: https://github.com/jqno/equalsverifier/compare/equalsverifier-3.18.1...equalsverifier-3.18.2
 [3.18.1]: https://github.com/jqno/equalsverifier/compare/equalsverifier-3.18...equalsverifier-3.18.1